+++ src/song/cdivisiontype.h
@@ -9,12 +9,12 @@ class CDivisionType
     : public QObject
 {
     Q_OBJECT
-    Q_PROPERTY(int enter READ enter WRITE enter NOTIFY enterChanged)
-    Q_PROPERTY(int time READ time WRITE time NOTIFY timeChanged)
+    Q_PROPERTY(qint16 enter READ enter WRITE enter NOTIFY enterChanged)
+    Q_PROPERTY(qint16 time READ time WRITE time NOTIFY timeChanged)
 
 public:
     explicit CDivisionType(QObject *parent = 0);
-    CDivisionType(int enter, int time, QObject *parent = 0);
+    CDivisionType(qint16 enter, qint16 time, QObject *parent = 0);
     CDivisionType(const CDivisionType &obj);
     CDivisionType& operator=(const CDivisionType &obj);
 
@@ -24,21 +24,21 @@ public:
     static const CDivisionType TRIPLET;
 
     // ---API--- //
-    inline int enter() const { return m_enter; }
-    inline int time()  const { return m_time; }
+    inline qint16 enter() const { return m_enter; }
+    inline qint16 time()  const { return m_time; }
 
-    inline void enter(const int val) { m_enter = val; emit enterChanged(); }
-    inline void time(const int val) { m_time = val; emit timeChanged(); }
+    inline void enter(const qint16 val) { m_enter = val; emit enterChanged(); }
+    inline void time(const qint16 val)  { m_time = val; emit timeChanged(); }
 
-    inline long convertTime(long in_time) const { return in_time * time() / enter(); }
+    inline qint64 convertTime(qint64 in_time) const { return in_time * time() / enter(); }
 
 signals:
     void enterChanged();
     void timeChanged();
 
 private:
-    int m_enter;
-    int m_time;
+    qint16 m_enter;
+    qint16 m_time;
 };
 
 #endif // CDIVISIONTYPE_H
